Undo use of key constants to access package
I thought it would be good practice to use string pseudo-constants to
define all the keys which are used to access the package dictionary
instead of hard coding them into the source code.

As it turns out, it does not make so much sense. The code readbility
drops a lot and it is very unlikely, that the key value will change in
the future. Typos should be catched with unit tests anyway.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>