Use absolute path for configuration file retrieval
The configuration file for the RLP CKAN harvester used a relative path
for reading the file, respectively only the filename itself. Since the
harvester classes extend a plugin class, they all will be instantiated
at the start of CKAN, independent of the fact, whether they are
activated as a plugin or not.

In this case the current working directoy is the one where CKAN was
launched and not the harvester directory. This commit fixes that by
using os.path.abspath.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>