Move harvester log filename into configuration
The log filename for the harvester should not be hard coded either. Thus
I have moved the log filename into the configuration file. As it seems
different harvester classes will make use of the URLs defined in the
configuration file. For that reason it might be best to use the
configuration reference on the global scope.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>