Add unit test for Berlin CKAN Harvester
I have created a new directory for unit tests. In the README file I have
appended a section describing a way to execute unit tests in the future.
To improve the testability I have refactored the code from the import
stage into using another method which is only responsible for ammending
the package.

Also with this commit I have moved the logger format code into the
configuration file.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>