Change ckanext-harvest dependency
I completed the feature that I wanted to implement on the feature branch
of our ckanext-harvest fork. This leads to using configuration in order
to import remote groups. As this is not yet reflected in the code here,
I have changed the pip requirement from using a certain branch to using
a certain commit where this implementation will keep working for now.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>