Add save object error call in case of exceptions
For the ckanext-harvest internal logging the extension uses methods to
store the error message in the database. This applies for the our case,
when the author fields are not set. Hence, I have added this call in the
try-except block. Also I have refactored the package change code for
Hamburg into a separate method as well.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>