Change default directory for the logfile
Until now there was no absolute path defined for the logfile, but a
relative path containing only the filename itself. Hence, the logfile
will be written to the place, where the CKAN process is started. This
can lead to access right problems.

For that reason the default path was changed to the directory where CKAN
log files are normally stored, unless it is installed from source.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>