Activate Moers Harvester plugin
One thing the implementation was missing is the generation of the unique
identifier since the datasets do not come with a predefined identifier.
It is generated the same way as in the other JSON dump harvester.

There is still a problem with the unicode characters in the name, I was
not able to replace them with their ASCII counterparts (ae, oe, ue, ss)
without decoding exception. This will be fixes as soon as possible.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>