Fix recursive log file directory creation
In case the log file path contains directories which do not exist yet I
had added an additional check to create the directories if necessary.

What I did not have in mind, was the fact that the method for creating
the directories should only take the directory path as argument not the
full path with the file name. For that reason the harvester created a
folder with the name govdata-harvester.log which lead to erroneous
behavior.

This commit fixes it by extracting the directory path before passing it
to the same methods.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>