Change revision for harvester in pip requirements
Since we are running CKAN 1.8 and the submitted pull requests were
merged into the ckanext-harvest master branch which is not working
anymore with CKAN 1.8 I have created a new branch.

In this branch, govdatade-compliant, I have cherry picked the features
and bug fixes I have submitted so we can use them as well.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>