Extend sector amendment test and implement patch
I have extended the unit test for the Berlin harvester's sector
amendment. This commit also implements the sector amendment.

From now on, the method amend_package should return a Boolean which
indicates whether the package is valid or not. In the latter case the
package is dropped. This way the code is a lot more testable.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>