Create Datahub IO CKAN harvester and first tests
There are now 4 packages which can be harvested from datahub.io. Hence,
a new harvester is created. Firstly, it will be tested, whether the
given packages are valid.

A package is valid, if it is part of a list of packages which should be
harvested. Secondly, the fetch stage will be overwritten to skip all
other packages.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>