Add normalize action dataset function
Obviously, the Model API /rest is too slow for remote validation. Hence,
I will use the Action API which then again needs normalization in order
to be validated against the JSON Schema. I have added this function and
the appropriate test cases. In addition, I had to extend the normalize
extras function so it normalizes string encoded strings, as well.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>