Implement functions to log URL failure and success
The functions provided by this commit encapsulate the logging of failure
and success of a URL. On failure a record is created or extended. When
the date has incremented by at least one day then the strike counter is
incremented as well.

The success function removes entries again. The next step is to combine
both methods in the actual link checking procedure.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>