Fix bugs in the link checker code
There were several problems with the link checker code. For one,
'deactivated' is not a valid dataset state. I had to extract the
actual date from both timestamps, otherwise time will be included.

Besides that I also noticed that socket.timeout is thrown, even though
it should have been wrapped inside of requests.exceptions.Timeout. Also
the RequestException is null. I have to figure out if it is still
possible to extract information in this case.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>